Chore : 생성자 추가
