+++ src/main/java/com/example/smartair/dto/roomDto/JoinRoomRequestDto.java
@@ -1,11 +1,11 @@
 package com.example.smartair.dto.roomDto;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Getter
 @Setter
+@Builder
+@AllArgsConstructor
 @NoArgsConstructor
 public class JoinRoomRequestDto {
     private String password;
